Make three reset glitch filter variants
The existing code for `resetGlitchFilter` was wrong: the `error` on
domains with unknown initial values rendered as an X value in HDL
instead of throwing an error.

The new `unsafeResetGlitchFilter` allows to sidestep the restriction if
the user is sure this works for them.

The new `resetGlitchFilterWithReset` is the better solution for domains
with unknown initial values, relying on an additional power-on reset
for which assertion glitches don't occur.